#11 сортировка товаров
