+++ src/components/showcase/Showcase.scss
@@ -10,35 +10,10 @@
     gap: 40px;
   }
 
-  &__sort-and-count {
-    display: flex;
-    justify-content: space-between;
-    gap: 30px;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 27px;
-  }
-
-  &__sort {
-    max-width: 262px;
-    width: 100%;
-  }
-
   &__products {
     display: flex;
     flex-wrap: wrap;
     gap: 50px 30px;
-  }
-
-  &__input {
-    color: rgba(0, 0, 0, 1);
-  }
-
-  &__bold {
-    font-size: 16px;
-    font-weight: 600;
-    line-height: 27px;
-    color: rgba(0, 0, 0, 1);
   }
 
   &__pagination {
