+++ src/components/sort/Sort.scss
@@ -1,0 +1,30 @@
+.sort {
+  max-width: 262px;
+  width: 100%;
+
+  &-and-count {
+    display: flex;
+    justify-content: space-between;
+    gap: 30px;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 27px;
+  }
+
+  &__input {
+    color: rgba(0, 0, 0, 1);
+  }
+
+  &__bold {
+    font-size: 16px;
+    font-weight: 600;
+    line-height: 27px;
+    color: rgba(0, 0, 0, 1);
+  }
+}
+
+
+
+
+
+
